+++ src/components/index.js
@@ -2,7 +2,6 @@ export { default as CBadge } from './badge';
 export { default as CBarCodeReader } from './barcodereader';
 export { default as CBigHeader } from './bigheader';
 export { default as CButton } from './button';
-export { default as CRescueButton } from './rescuebutton';
 export { default as CContent } from './content';
 export { default as CError } from './error';
 export { default as CGreyBox } from './greybox';
@@ -11,6 +10,7 @@ export { default as CImage } from './image';
 export { default as CInput } from './input';
 export { default as CMiniHeader } from './miniheader';
 export { default as CModal } from './modal';
+export { default as CRescueButton } from './rescuebutton';
 export { default as CSep } from './sep';
 export { default as CSpace } from './space';
 export { default as CSpinner } from './spinner';
@@ -20,8 +20,10 @@ export { default as CWaypointAddress } from './wpaddress';
 export { default as CWaypointButtons } from './wpbuttons';
 export { default as CWaypointCounters } from './wpcounters';
 export { default as CWaypointOtherPassage } from './wpotherpassage';
+export { default as CWaypointTemplate } from './wptemplate';
 
 export const DEFAULT_FONT_SIZE = 15;
+
 export const COLORS = {
   WHITE: '#ffffff',
   BLACK: '#333333',