+++ src/screens/index.js
@@ -1,9 +1,10 @@
-export { default as ScreenSplash } from './splash';
-export { default as ScreenDriver } from './driver';
 export { default as ScreenCar } from './car';
+export { default as ScreenDriver } from './driver';
 export { default as ScreenManagers } from './managers';
-export { default as ScreenWaypointDashboard } from './wpdashboard';
+export { default as ScreenSplash } from './splash';
+export { default as ScreenWaypointBadCondition } from './wpbadcondition';
 export { default as ScreenWaypointCollection } from './wpcollection';
+export { default as ScreenWaypointDashboard } from './wpdashboard';
 export { default as ScreenWaypointGalery } from './wpgalery';
 
 export const NAVS = {